removed old tf distributed example
